Permissão

+++ app/Http/Controllers/ProjectController.php
@@ -19,7 +19,7 @@ class ProjectController extends Controller
 
     public function index()
     {
-        return $this->repository->all();
+        return $this->repository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     public function store(Request $request)
@@ -29,16 +29,52 @@ class ProjectController extends Controller
 
     public function show($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+
         return $this->repository->find($id);
     }
 
     public function destroy($id)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+
         $this->repository->delete($id);
     }
 
     public function update($id, Request $request)
     {
+        if ($this->checkProjectPermissions($id) == false) {
+            return ['error' => 'Access forbidden'];
+        }
+
         return $this->service->update($request->all(), $id);
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->repository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) || $this->checkProjectMember($projectId)) {
+            return true;
+        }
+
+        return false;
+
     }
 }
